chore: move register.jsx to src
